Move stories to sit with components
